Initialize a flash factor using IRLBA

Initializes a new ("greedy") flash factor using irlba. This can be somewhat faster than flash_greedy_init_default for large, dense data matrices. For sparse matrices of class Matrix, the default initialization should generally be preferred.

flash_greedy_init_irlba(flash, seed = 666, ...)

Arguments

flash

A flash_fit object.

seed

Since initialization is random, a default seed is set for reproducibility.

...

Additional parameters to be passed to irlba.

Value

A list of length two consisting of, respectively, the vector of initial values for loadings \(\ell_{\cdot k}\) and the vector of initial factor values \(f_{\cdot k}\).

See also