GTEx data

Derived from data made available by the Genotype Tissue Expression (GTEx) project (Lonsdale et al. 2013), which provides \(z\)-scores for assessing the significance of effects of genetic variants (single nucleotide polymorphisms, or SNPs) on gene expression across 44 human tissues. To reduce the data to a more manageable size, Urbut et al. (2019) chose the "top" SNP for each gene --- that is, the SNP associated with the largest (absolute) \(z\)-score over all 44 tissues. This yields a \(16,069 \times 44\) matrix of \(z\)-scores, with rows corresponding to SNP-gene pairs and columns corresponding to tissues. The dataset included here is further subsampled down to 1000 rows.

Format

gtex is a matrix with 1000 rows and 44 columns, with rows corresponding to SNP-gene pairs and columns corresponding to tissues.

Source

https://github.com/stephenslab/gtexresults/blob/master/data/MatrixEQTLSumStats.Portable.Z.rds

References

Lonsdale et al. (2013). "The Genotype-Tissue Expression (GTEx) project." Nature Genetics 45(6), 580--585.

Urbut, Wang, Carbonetto, and Stephens (2019). "Flexible statistical methods for estimating and testing effects in genomic studies with multiple conditions." Nature Genetics 51(1), 187--195.

Examples

data(gtex)
summary(gtex)
#>  Adipose_Subcutaneous Adipose_Visceral_Omentum Adrenal_Gland      
#>  Min.   :-22.19423    Min.   :-14.87582        Min.   :-12.57593  
#>  1st Qu.: -1.84280    1st Qu.: -1.57162        1st Qu.: -1.26369  
#>  Median : -0.06988    Median : -0.05287        Median : -0.01588  
#>  Mean   : -0.27107    Mean   : -0.16151        Mean   : -0.16142  
#>  3rd Qu.:  1.48032    3rd Qu.:  1.19198        3rd Qu.:  0.93398  
#>  Max.   : 21.08357    Max.   : 16.43779        Max.   : 13.65782  
#>   Artery_Aorta       Artery_Coronary     Artery_Tibial       
#>  Min.   :-17.62629   Min.   :-12.74545   Min.   :-23.307430  
#>  1st Qu.: -1.45706   1st Qu.: -1.11132   1st Qu.: -1.598985  
#>  Median : -0.01537   Median : -0.06018   Median :  0.008435  
#>  Mean   : -0.15449   Mean   : -0.11235   Mean   : -0.134894  
#>  3rd Qu.:  1.27376   3rd Qu.:  1.03352   3rd Qu.:  1.497730  
#>  Max.   : 18.06622   Max.   : 11.14108   Max.   : 22.814513  
#>  Brain_Anterior_cingulate_cortex_BA24 Brain_Caudate_basal_ganglia
#>  Min.   :-8.99940                     Min.   :-12.12301          
#>  1st Qu.:-1.08092                     1st Qu.: -1.16093          
#>  Median :-0.03334                     Median :  0.05487          
#>  Mean   :-0.07624                     Mean   :  0.00441          
#>  3rd Qu.: 0.96140                     3rd Qu.:  1.06846          
#>  Max.   : 7.90574                     Max.   : 11.00090          
#>  Brain_Cerebellar_Hemisphere Brain_Cerebellum     Brain_Cortex      
#>  Min.   :-12.144353          Min.   :-12.86501   Min.   :-10.85760  
#>  1st Qu.: -1.098736          1st Qu.: -1.19935   1st Qu.: -1.14260  
#>  Median :  0.020778          Median : -0.03972   Median : -0.03468  
#>  Mean   :  0.000413          Mean   : -0.01722   Mean   : -0.05909  
#>  3rd Qu.:  1.107743          3rd Qu.:  1.10800   3rd Qu.:  1.01669  
#>  Max.   : 10.897779          Max.   : 11.67528   Max.   : 11.59038  
#>  Brain_Frontal_Cortex_BA9 Brain_Hippocampus    Brain_Hypothalamus  
#>  Min.   :-11.80390        Min.   :-11.232401   Min.   :-10.031394  
#>  1st Qu.: -1.14363        1st Qu.: -1.023547   1st Qu.: -0.905895  
#>  Median : -0.07074        Median :  0.005028   Median : -0.007766  
#>  Mean   : -0.05440        Mean   :  0.005987   Mean   : -0.043437  
#>  3rd Qu.:  1.02602        3rd Qu.:  1.007408   3rd Qu.:  0.879970  
#>  Max.   : 10.35621        Max.   :  8.075452   Max.   :  8.981126  
#>  Brain_Nucleus_accumbens_basal_ganglia Brain_Putamen_basal_ganglia
#>  Min.   :-10.8223                      Min.   :-9.46936           
#>  1st Qu.: -1.1396                      1st Qu.:-1.10615           
#>  Median : -0.1130                      Median :-0.04910           
#>  Mean   : -0.1048                      Mean   :-0.06563           
#>  3rd Qu.:  0.9185                      3rd Qu.: 1.00442           
#>  Max.   :  9.5235                      Max.   : 8.90531           
#>  Breast_Mammary_Tissue Cells_EBV-transformed_lymphocytes
#>  Min.   :-15.6587      Min.   :-10.57911                
#>  1st Qu.: -1.4741      1st Qu.: -1.25969                
#>  Median : -0.0223      Median : -0.03338                
#>  Mean   : -0.1313      Mean   : -0.10913                
#>  3rd Qu.:  1.2819      3rd Qu.:  1.02339                
#>  Max.   : 16.8237      Max.   : 14.77698                
#>  Cells_Transformed_fibroblasts Colon_Sigmoid       Colon_Transverse   
#>  Min.   :-17.0952              Min.   :-13.33327   Min.   :-15.36002  
#>  1st Qu.: -1.7708              1st Qu.: -1.34944   1st Qu.: -1.36955  
#>  Median : -0.1495              Median : -0.06941   Median : -0.03634  
#>  Mean   : -0.3416              Mean   : -0.16364   Mean   : -0.20777  
#>  3rd Qu.:  1.2835              3rd Qu.:  1.04081   3rd Qu.:  1.06871  
#>  Max.   : 19.2887              Max.   : 13.37116   Max.   : 15.64862  
#>  Esophagus_Gastroesophageal_Junction Esophagus_Mucosa   Esophagus_Muscularis
#>  Min.   :-13.42037                   Min.   :-19.5164   Min.   :-18.61334   
#>  1st Qu.: -1.27899                   1st Qu.: -1.5160   1st Qu.: -1.71632   
#>  Median : -0.03443                   Median : -0.1147   Median : -0.06828   
#>  Mean   : -0.13436                   Mean   : -0.1280   Mean   : -0.25893   
#>  3rd Qu.:  1.05542                   3rd Qu.:  1.3108   3rd Qu.:  1.35094   
#>  Max.   : 12.27858                   Max.   : 19.6146   Max.   : 18.45721   
#>  Heart_Atrial_Appendage Heart_Left_Ventricle     Liver           
#>  Min.   :-14.16783      Min.   :-15.2517     Min.   :-11.116238  
#>  1st Qu.: -1.45627      1st Qu.: -1.3879     1st Qu.: -0.921782  
#>  Median : -0.07628      Median : -0.0378     Median :  0.007326  
#>  Mean   : -0.18849      Mean   : -0.2160     Mean   : -0.012890  
#>  3rd Qu.:  1.10405      3rd Qu.:  1.0856     3rd Qu.:  0.965711  
#>  Max.   : 14.05779      Max.   : 15.3510     Max.   :  8.653149  
#>       Lung          Muscle_Skeletal     Nerve_Tibial           Ovary          
#>  Min.   :-20.4296   Min.   :-23.1101   Min.   :-20.97164   Min.   :-10.31093  
#>  1st Qu.: -1.6291   1st Qu.: -1.7727   1st Qu.: -1.88211   1st Qu.: -1.10607  
#>  Median : -0.1825   Median : -0.1185   Median : -0.02156   Median :  0.02842  
#>  Mean   : -0.2045   Mean   : -0.3323   Mean   : -0.21575   Mean   : -0.05246  
#>  3rd Qu.:  1.3875   3rd Qu.:  1.3392   3rd Qu.:  1.52011   3rd Qu.:  1.02720  
#>  Max.   : 21.4307   Max.   : 20.0308   Max.   : 22.36088   Max.   : 10.35619  
#>     Pancreas           Pituitary           Prostate        
#>  Min.   :-13.85419   Min.   :-11.0648   Min.   :-10.15243  
#>  1st Qu.: -1.36881   1st Qu.: -1.2287   1st Qu.: -1.13629  
#>  Median : -0.07797   Median : -0.1005   Median : -0.12194  
#>  Mean   : -0.17514   Mean   : -0.1267   Mean   : -0.07537  
#>  3rd Qu.:  1.10445   3rd Qu.:  1.0129   3rd Qu.:  0.97536  
#>  Max.   : 15.43049   Max.   : 10.3997   Max.   : 10.83603  
#>  Skin_Not_Sun_Exposed_Suprapubic Skin_Sun_Exposed_Lower_leg
#>  Min.   :-16.80919               Min.   :-23.10359         
#>  1st Qu.: -1.41420               1st Qu.: -1.72428         
#>  Median :  0.03397               Median : -0.08678         
#>  Mean   : -0.04676               Mean   : -0.15531         
#>  3rd Qu.:  1.30838               3rd Qu.:  1.41275         
#>  Max.   : 15.69632               Max.   : 22.23993         
#>  Small_Intestine_Terminal_Ileum     Spleen            Stomach         
#>  Min.   :-8.719451              Min.   :-10.6512   Min.   :-14.36215  
#>  1st Qu.:-1.079962              1st Qu.: -1.2765   1st Qu.: -1.35343  
#>  Median : 0.001343              Median : -0.1641   Median : -0.05671  
#>  Mean   :-0.102133              Mean   : -0.1566   Mean   : -0.17118  
#>  3rd Qu.: 0.971632              3rd Qu.:  0.9506   3rd Qu.:  1.19389  
#>  Max.   : 8.807607              Max.   : 11.4742   Max.   : 14.75691  
#>      Testis            Thyroid              Uterus             Vagina        
#>  Min.   :-12.4311   Min.   :-20.74476   Min.   :-8.33238   Min.   :-8.60648  
#>  1st Qu.: -1.4523   1st Qu.: -1.81461   1st Qu.:-1.12685   1st Qu.:-1.04414  
#>  Median : -0.1939   Median : -0.01762   Median :-0.01644   Median :-0.02833  
#>  Mean   : -0.1863   Mean   : -0.11999   Mean   :-0.10482   Mean   :-0.05341  
#>  3rd Qu.:  1.1162   3rd Qu.:  1.55465   3rd Qu.: 0.87098   3rd Qu.: 0.96256  
#>  Max.   : 16.4308   Max.   : 22.27820   Max.   : 9.66800   Max.   : 9.45592  
#>   Whole_Blood      
#>  Min.   :-23.1985  
#>  1st Qu.: -1.4820  
#>  Median : -0.1303  
#>  Mean   : -0.1078  
#>  3rd Qu.:  1.1960  
#>  Max.   : 19.4720